Moved all Resource I/O from OGFILE2.cpp to OGFILE_RES.cpp.
