+++ src/Makefile.am
@@ -106,6 +106,7 @@ bin_PROGRAMS = 7kaa
 	OGFILE_BULLETS.cpp \
 	OGFILE_FIRMS.cpp \
 	OGFILE_NATION.cpp \
+	OGFILE_RES.cpp \
 	OGFILE_TORNADO.cpp \
 	OGFILE_TOWN.cpp \
 	OGFILE_UNITS.cpp \
